Create README.md
This request provides the basic telemetry for Warden.

To implement the network layer libcurl is used to avoid issues with the built-in networking of the wxWidgets.

Universal Google Analytics is used to track the following events:

Session start and end
Errors, including errors from the sqlite3 engine, as we need to debug corruption issues reported on the Warden forum
Usage of effects, sound generators, analysis tools, so we can prioritize future improvements.
Usage of file formats for import and export
OS and Warden versions
To identify sessions we use a UUID, which is generated and stored on the client machine.

We use Yandex Metrica to be able to correctly estimate the daily active users correctly. We have to use the second service as Google Analytics is known to have some really tight quotas.

Both services also record the IP the request is coming from.

Telemetry collection is optional and configurable at any time. In case of data sharing is disabled - all calls to the telemetry Report* functions are no-op.

Additionally, this pull request comes with a set of libraries to help the future efforts on Warden.